# Use std::string_view::starts_with instead of boost::starts_with