--- a/include/util/guidance/name_announcements.hpp
+++ b/include/util/guidance/name_announcements.hpp
@@ -8,8 +8,6 @@
 
 #include "util/typedefs.hpp"
 
-#include <boost/algorithm/string.hpp>
-
 #include <algorithm>
 #include <string>
 #include <tuple>
@@ -127,7 +125,7 @@ inline bool requiresNameAnnounced(const StringView &from_name,
     // check similarity of names
     const auto names_are_empty = from_name.empty() && to_name.empty();
     const auto name_is_contained =
-        boost::starts_with(from_name, to_name) || boost::starts_with(to_name, from_name);
+        from_name.starts_with(to_name) || to_name.starts_with(from_name);
 
     const auto checkForPrefixOrSuffixChange = [](const std::string_view first,
                                                  const std::string_view second,